pyproject.toml
[tool.poetry]
name = "Assessment"
version = "0.1.0"
description = "description"
authors = ["authors"]
license = "proprietary"
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "~3.12"
uvicorn = { extras = ["standard"], version = "^0.24.0.post1" }
fastapi = "^0.104.1"
fastapi-injector = "^0.5.3"
alembic = "^1.12.1"
asyncpg = "^0.29.0"
cryptography = "^41.0.5"
greenlet = "^3.0.1"
injector = "^0.21.0"
cachetools = "^5.3.3"
python-jose = { version = "^3.3.0", extras = ["cryptography"] }
pyjwt = "^2.8.0"
sqlalchemy = "^2.0.23"
pydantic = {extras = ["email"], version = "^2.5.2"}
APScheduler = "^3.10.4"
requests = "^2.31.0"
types-requests = "^2.31.0.10"

[tool.poetry.dev-dependencies]
mypy = "^1.7.1"
black = "^23.11.0"
flake8 = "^6.1.0"
flake8-bugbear = "^23.9.16"
flake8-builtins = "^2.2.0"
flake8-class-attributes-order = "^0.1.3"
flake8-warnings = "^0.4.1"


[tool.poetry.group.dev.dependencies]
mypy = "^1.7.1"
black = "^23.11.0"
flake8 = "^6.1.0"
flake8-bugbear = "^23.9.16"
flake8-builtins = "^2.2.0"
flake8-class-attributes-order = "^0.1.3"
flake8-warnings = "^0.4.1"

[tool.mypy]
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = ["asyncpg.*"]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
src = "src.__main__:main"
